+++ src/sdk/__tests__/getAuthenticatedAccount.test.ts
@@ -1,0 +1,25 @@
+import nock from 'nock';
+import { getAuthenticatedAccount } from '..';
+
+describe('getAuthenticatedAccount', () => {
+  it('should return an Account', async () => {
+    nock('https://permanent.local')
+      .get('/api/account/getAuthenticatedAccount')
+      .replyWithFile(
+        200,
+        `${__dirname}/fixtures/getAuthenticatedAccount/accountVo.json`,
+        {
+          'Content-Type': 'application/json',
+        },
+      );
+
+    const account = await getAuthenticatedAccount(
+      {
+        bearerToken: '12345',
+        baseUrl: 'https://permanent.local/api',
+      },
+    );
+
+    expect(account).toMatchSnapshot();
+  });
+});